Início do fluxo de adição de categorias ao usuário

+++ src/main/java/com/integrador/entities/Categoria.java
@@ -23,7 +23,7 @@ public class Categoria implements Serializable {
 	@Id
 	@Column(name = "CATEGORIA_ID")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	@Column(name = "TITULO")
 	private String titulo;
 
@@ -38,17 +38,17 @@ public class Categoria implements Serializable {
 	public Categoria() {
 	}
 
-	public Categoria(Integer id, String titulo) {
+	public Categoria(Long id, String titulo) {
 		super();
 		this.id = id;
 		this.titulo = titulo;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
